refactor: split TUI into focused modules
