+++ rclip/tui/__init__.py
@@ -1,0 +1,3 @@
+from rclip.tui.app import run_tui
+
+__all__ = ["run_tui"]
